refractored code

+++ app/routers/logs.py
@@ -1,26 +1,26 @@
 # dependancies:
 import datetime as dt
 
+from core.auth import AuthHandler
 from core.models import Logs as ModelLogs
 from core.schema import Log as SchemaLog
 from db.database import SessionLocal
-from fastapi import APIRouter, HTTPException, Query, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from psycopg2.errors import UniqueViolation
 
 # error handling
 from sqlalchemy.exc import IntegrityError
 
-from routers.helperfunctions import convertWKBtoWKT
+from routers.helpers.authSharedFunctions import checkRoleAdmin
 
 logsRouter = APIRouter()
 
 db = SessionLocal()
-
+auth_handler = AuthHandler()
 
 #################################################################################################################################
 #                                                  Create                                                                       #
 #################################################################################################################################
-@logsRouter.post("/create", response_model=SchemaLog)
 def add_log(log: SchemaLog):
     try:
         log = ModelLogs(date=dt.datetime.today(), data=log.log_data)
@@ -36,7 +36,10 @@ def add_log(log: SchemaLog):
 #                                                  Read                                                                         #
 #################################################################################################################################
 @logsRouter.get("/read")
-def get_log():
+def get_log(payload=Depends(auth_handler.auth_wrapper)):
+
+    if checkRoleAdmin(payload) == False:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authorized")
     try:
         result = db.query(ModelLogs).all()
     except Exception:
@@ -45,7 +48,11 @@ def get_log():
 
 
 @logsRouter.get("/read/{log_date}")
-def get_log(log_date: str):
+def get_log(log_date: str, payload=Depends(auth_handler.auth_wrapper)):
+
+    if checkRoleAdmin(payload) == False:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authorized")
+
     try:
         result = db.query(ModelLogs).filter(ModelLogs.date == log_date).first()
     except Exception:
@@ -57,12 +64,27 @@ def get_log(log_date: str):
 #################################################################################################################################
 #                                                  Update                                                                       #
 #################################################################################################################################
+# @logsRouter.get("test")
+# def test():
+#     payload = {"role": "admin"}
+#     return authtest(payload)
+
+
+# def authtest(payload=Depends(auth_handler.auth_wrapper)):
+#     if checkRoleAdmin(payload) == False:
+#         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authorized")
+#     return "success"
+
 
 #################################################################################################################################
 #                                                  Delete                                                                       #
 #################################################################################################################################
 @logsRouter.delete("/delete/{log_date}")
-def delete_log(log_date: str):
+def delete_log(log_date: str, payload=Depends(auth_handler.auth_wrapper)):
+
+    if checkRoleAdmin(payload) == False:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authorized")
+
     try:
         db.query(ModelLogs).filter(ModelLogs.date == log_date).delete()
         db.commit()
